Move space sync into helper
